Add async version of driver (hidden behind async feature gate)
